requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
seaborn==0.13.2
uvicorn==0.34.0
pandas>=1.0.0
typer[all]>=0.9.0
scikit-learn>=0.24.0
transformers>=4.0.0
fastapi>==0.95.2
pydantic>==1.10.8
indently>==1.0.0
pytorch-lightning>==2.0.0
loguru==0.7.3
python-dotenv>=0.7.0
wandb>=0.10.0
